Revert "net/lwip: use sem_tickwait() instead of sem_timedwait()"
This reverts commit c07e728debf3549fae9522252e0a142cc310419f.

This commit was merged to improve network bandwidth, but it breaks
a behavior expected by some other pieces of code, such as the
webserver teardown mechanism. The issue when this patch is applied
affects all lwip functions relying on semaphores with timeouts, causing
the semaphore wait function to remain blocked even when the timeout
occurs.

Change-Id: I8dfd056a8e3811eb7e0174ec53183f0a603cbe81
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>