netutils/dhcpd: fix unexpected reset of the network configuration
Stopping the DHCP server is asynchronous and actual release happens
whithin 60 seconds. This causes issues to applications that stop the
DHCP server then change the network configuration before the DHCP
is actually stopped. As a result, the new network configuration is
reset by the DHCP server stop mechanism. This patch disables the
network configuration reset, but should be later replaced by a
cleaner and smarter mechanism.

Change-Id: I7a61ef4a00913ded4883969aae9bf940e6c1f1f7
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>